fix: bug

+++ src/freqfine.rs
@@ -50,7 +50,7 @@ impl FreqFine {
         let actual: u8 = self.read(i2c, self.address, ADDR)?;
         
         const ODR_COEFF: u16 = 64
-        let ODR_actual: f32 = 6667 + (0.0015 * self.value * 6667 / ODR_COEFF)
+        let ODR_actual: f32 = 6667 + (0.0015 * (self.value as f32) * 6667 / ODR_COEFF)
         
         Ok(ODR_actual)
     }
